move launch code to own file
