+++ harpy/deconvolve.py
@@ -6,7 +6,8 @@ from rich import box
 from rich.table import Table
 import rich_click as click
 from ._conda import generate_conda_deps
-from ._misc import fetch_rule, snakemake_log, launch_snakemake
+from ._launch import launch_snakemake
+from ._misc import fetch_rule, snakemake_log
 from ._parsers import parse_fastq_inputs
 
 docstring = {
