+++ harpy/impute.py
@@ -7,7 +7,8 @@ from rich import box
 from rich.table import Table
 import rich_click as click
 from ._conda import generate_conda_deps
-from ._misc import fetch_rule, fetch_report, fetch_script, snakemake_log, launch_snakemake
+from ._launch import launch_snakemake
+from ._misc import fetch_rule, fetch_report, fetch_script, snakemake_log
 from ._parsers import parse_alignment_inputs, biallelic_contigs
 from ._validations import validate_input_by_ext, vcf_samplematch, check_impute_params, validate_bam_RG
