+++ harpy/resume.py
@@ -8,7 +8,8 @@ from rich.table import Table
 import rich_click as click
 from ._validations import check_envdir
 from ._printing import print_error
-from ._misc import snakemake_log, launch_snakemake
+from ._launch import launch_snakemake
+from ._misc import snakemake_log
 from ._conda import generate_conda_deps
 
 @click.command(no_args_is_help = True, epilog = "See the documentation for more information: https://pdimens.github.io/harpy/modules/other")
