+++ harpy/simulate.py
@@ -6,7 +6,8 @@ from rich import box
 from rich.table import Table
 import rich_click as click
 from ._conda import generate_conda_deps
-from ._misc import fetch_rule, fetch_script, snakemake_log, launch_snakemake
+from ._launch import launch_snakemake
+from ._misc import fetch_rule, fetch_script, snakemake_log
 from ._printing import print_error
 from ._validations import validate_input_by_ext, check_fasta
